consolidate ProviderRollup under components/filters
